Fixed issue with Arista config data not making it into priciple.
Since the neutron-api charm cannot accept more than one additional
config file, we place all the Arista configuration information in
ml2_conf.ini. We programatically duplicate this information into
ml2_conf_arista.ini, so that Arista has it in the expected place, if
need be.

Also added a region-name param to the config.yaml, which gets written
out to the .ini.